added car icon and equivalent driven km in chart views

+++ app/src/main/java/wamboo/example/videocompressor/models/CompressData.kt
@@ -7,7 +7,7 @@ import java.math.BigDecimal
 @Entity(tableName = "compression")
 class CompressData(
     val sizeReduction: Long,
-    val co2: Int,
+    val co2: Long,
     val milliSeconds: Long,
     val date: String,
     @PrimaryKey(autoGenerate = true)
